Refactor: 피드 목록 리팩토링
